machine-005: rename getcookie
